Added a second layer, feature regularization, feature scaling, learning curves, a gradient, descent implementation, and a back propagation implementation check by computing the numerical gradient
In this iteration of the code I added a second layer to my neural network and also added feature scaling and feature regularization to help increase the ability of the algorithm to accurately identify numbers. I also added code to compute the numerical gradient as a means to check if the backpropagation algorithm was implemented correctly, and I also added the ability to view the leraning curve of the algorithm. Finally I cleaned up some code and to use more of the optimized scipy library, and some numpy array properties to make the code look (slightly) cleaner